pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Craxell <andresaltamirasanz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.115.0"}
pymongo = "^4.8.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
pydantic-settings = "^2.5.2"
python-dotenv = "^1.0.1"
bcrypt = "^4.2.0"
chromadb = "^0.5.7"
coverage = "^7.6.1"
tiktoken = "^0.8.0"
pypdf2 = "^3.0.1"
python-docx = "^1.1.2"
openai = "0.28"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
